requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-nested-admin==4.1.1
Pillow==11.0.0
python-dotenv==1.0.0
requests==2.31.0
psycopg==3.1.14
dj-database-url==2.1.0
gunicorn==21.2.0
reportlab==4.0.9